# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:63531",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/products/{id}/publish" : {
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "publish",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PublishClaimRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/attributes" : {
      "post" : {
        "tags" : [ "attribute-controller" ],
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAttributeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attributes/{key}/values" : {
      "post" : {
        "tags" : [ "attribute-controller" ],
        "operationId" : "addValue",
        "parameters" : [ {
          "name" : "key",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddEnumValueRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attribute-schemas/{id}/fields" : {
      "post" : {
        "tags" : [ "attribute-controller" ],
        "operationId" : "addField",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddSchemaFieldRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/attributes/{key}" : {
      "get" : {
        "tags" : [ "attribute-controller" ],
        "operationId" : "getAttribute",
        "parameters" : [ {
          "name" : "key",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AttributeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attribute-schemas/{id}" : {
      "get" : {
        "tags" : [ "attribute-controller" ],
        "operationId" : "getSchema",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SchemaResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "IdResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ProductResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "productType" : {
            "type" : "string"
          },
          "lifecycle" : {
            "type" : "string"
          },
          "brandCode" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "classification" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "taxonomyPath" : {
            "type" : "string"
          }
        }
      },
...
      "PublishClaimRequest" : {
        "type" : "object",
        "properties" : {
          "attributeKey" : {
            "type" : "string"
          },
          "evidenceRefs" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "CreateAttributeRequest" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "governance" : {
            "type" : "string"
          },
          "knownValues" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "AddEnumValueRequest" : {
        "type" : "object",
        "properties" : {
          "value" : {
            "type" : "string"
          }
        }
      },
      "AddSchemaFieldRequest" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "required" : {
            "type" : "boolean"
          },
          "allowBreaking" : {
            "type" : "boolean"
          }
        }
      },
...
      "AttributeResponse" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "string"
          },
          "governance" : {
            "type" : "string"
          },
          "knownValues" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "status" : {
            "type" : "string"
          }
        }
      },
      "SchemaFieldResponse" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "required" : {
            "type" : "boolean"
          }
        }
      },
      "SchemaResponse" : {
        "type" : "object",
        "properties" : {
          "schemaId" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "scope" : {
            "type" : "string"
          },
          "fields" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SchemaFieldResponse"
            }
          },
          "status" : {
            "type" : "string"
          }
        }
      }
    }
  }
}